seperate files for Array objects

+++ arrayworkflow.py
@@ -220,7 +220,7 @@ class Array:
 		afile.close()
 		self.dic = {}
 		for l in alines[1:]:
-			self.dic[l.split('\t')[0]] = int(l.split('\t')[1].strip('\n'))
+			self.dic[l.split('\t')[0]] = float(l.split('\t')[1].strip('\n'))
 		self.snps = self.dic.keys()
